Delayed assignment cleanup
Immediately deleting role assignments is undesirable sometimes. People
might forget to renew their ldap group membership, be on vacation or
have other temporary short-term leave. If the role assignment is gone,
projects are left unattended, and it is not desirable.

Modify the assignment script to allow for certain time period, when the
orphaned assignment should not be deleted. The timestamps are stored in
redis.